Cosmetics
Split PlayResult into PlayResult + Checker;
Added Checker type to package object;
Removed type Named.
