+++ src/main/scala/com/phasmidsoftware/bridge/director/package.scala
@@ -4,6 +4,11 @@
 
 package com.phasmidsoftware.bridge
 
+import com.phasmidsoftware.misc.JPredicate
+
 package object director {
   type Pos = (Int, Card) // XXX Pair number and their "card"
+
+  type Checker = JPredicate[ScoreVul]
+
 }
